Create Stationarity Test
A time series is said to be weakly stationary if it has constant mean and variance, and the covariance between any two observations depends on the lag and not time. The Zivot- Andrew (1992) test can be used to test for stationary in the data. It also accounts for structural breaks in the data which could be due to significant economic events, political changes or external shocks.